Fix `blurb` import in the runpy interface
When `blurb` was being restructured, the import paths moved to deeper levels of the project layout. It was updated in the console scripts so invoking just the `blurb` command worked but `python -m blurb` didn't. This resulted in breaking the MSI installer smoke testing workflow in CPython's CI [[1]].

Blurb's own CI never invokes said interface, so it had no chance of catching the regression.

This patch corrects the import but does not make an attempt to improve testing.

[1]: https://github.com/python/cpython/issues/121879